chore: dev mode check to shutdown function
A dev mode condition check is added to the shutdown function in `shutdown.js`. This adjustment allows the shutdown process to bypass during development for testing, avoiding unnecessary disruptions. A log statement has also been added within the shutdown function for better tracking.

+++ src/main/shutdown.js
@@ -1,8 +1,13 @@
 import { exec } from 'child_process';
 import { getEventSender, initEventSender } from './event-sender';
 import { delay } from '../utils/delay';
+import { is } from '@electron-toolkit/utils';
 
 function shutdownComputer(log) {
+  if (is.dev) {
+    log('Dev Mode: Would have shut down the computer now');
+    return;
+  }
   delay(5000).then(() => {
     log('Gute Nacht! Shutting down...');
     exec('shutdown /s /f /t 0', (error, stdout, stderr) => {
@@ -17,6 +22,7 @@ function shutdownComputer(log) {
 }
 
 export function shutdown(event) {
+  console.log('Shutting down...');
   initEventSender(event);
   shutdownComputer(getEventSender());
 }
